+++ HotEvents/app/src/main/java/com/example/hotevents/EventDetailsActivity.java
@@ -21,15 +21,27 @@ import android.widget.ImageView;
 import android.widget.TextView;
 import android.widget.Toast;
 
+import com.google.android.gms.tasks.OnCompleteListener;
 import com.google.android.gms.tasks.OnFailureListener;
 import com.google.android.gms.tasks.OnSuccessListener;
+import com.google.android.gms.tasks.Task;
 import com.google.android.material.tabs.TabLayout;
 import com.google.firebase.firestore.DocumentSnapshot;
 import com.google.firebase.firestore.FieldValue;
 import com.google.firebase.firestore.FirebaseFirestore;
 
+import org.json.JSONException;
+import org.json.JSONObject;
+
+import java.io.IOException;
 import java.util.HashMap;
 import java.util.Map;
+
+import okhttp3.MediaType;
+import okhttp3.OkHttpClient;
+import okhttp3.Request;
+import okhttp3.RequestBody;
+import okhttp3.Response;
 
 public class EventDetailsActivity extends AppCompatActivity {
 
@@ -43,6 +55,11 @@ public class EventDetailsActivity extends AppCompatActivity {
     ViewPager2 viewPager2;
     EventPagerAdapter eventPagerAdapter;
     TextView organiserName;
+    String eventId;
+    String myeventTitle;
+    String organizerId;
+    String orgfcmToken;
+    private FirebaseFirestore db;
     private static final String TAG = "EventDetailsActivity";
 
     @Override
@@ -57,6 +74,17 @@ public class EventDetailsActivity extends AppCompatActivity {
 
         if (myEvent.getTitle() != null) {
             eventTitle.setText(myEvent.getTitle());
+        }
+        if (myEvent.getTitle() != null) {
+            myeventTitle = myEvent.getTitle();
+        }
+
+        if (myEvent.getEventId() != null) {
+            eventId = myEvent.getEventId();
+        }
+
+        if (myEvent.getOrganiserId() != null) {
+             organizerId = myEvent.getOrganiserId();
         }
 
 
@@ -106,6 +134,7 @@ public class EventDetailsActivity extends AppCompatActivity {
             }
         });
 
+
         Button signUpButton = findViewById(R.id.check_in_button);
         signUpButton.setOnClickListener(new View.OnClickListener() {
             @Override
@@ -116,10 +145,42 @@ public class EventDetailsActivity extends AppCompatActivity {
                         requestPermissions(new String[]{Manifest.permission.POST_NOTIFICATIONS}, 1);
                     }
                 }
+                Log.d(TAG, "Attempting to send milestone notifications");
+                fetchFCMTokenForOrganizer(organizerId);
             }
         });
-
     }
+
+    public void fetchFCMTokenForOrganizer(String organizerId) {
+        db = FirebaseFirestore.getInstance();
+        db.collection("Users").document(organizerId).get()
+                .addOnCompleteListener(task -> {
+                    if (task.isSuccessful()) {
+                        DocumentSnapshot document = task.getResult();
+                        if (document.exists()) {
+                            orgfcmToken = document.getString("fcmToken");
+                            if (orgfcmToken != null) {
+                                // FCM token found, you can use it here
+                                sendMilestoneIfConditionMet();
+                                Log.d(TAG, "FCM Token for organizer " + organizerId + ": " + orgfcmToken);
+                                // Now you can perform any actions with the FCM token
+                            } else {
+                                // FCM token not found for the organizer
+                                Log.d(TAG, "FCM token not found for organizer ID: " + organizerId);
+                            }
+                        } else {
+                            // Document not found
+                            Log.d(TAG, "No document found for organizer ID: " + organizerId);
+                        }
+                    } else {
+                        // Error fetching document
+                        Log.e(TAG, "Error fetching document: ", task.getException());
+                    }
+                });
+    }
+
+
+
     /**
      * Handles the sign-up button click event.
      * Retrieves the device ID of the user and the FCM token from Firestore Users collection.
@@ -180,6 +241,78 @@ public class EventDetailsActivity extends AppCompatActivity {
                 .addOnFailureListener(e -> {
                     Log.e(TAG, "Error updating mysignup array", e);
                 });
+    }
+
+    void sendMilestoneIfConditionMet() {
+        FirebaseFirestore db = FirebaseFirestore.getInstance();
+        db.collection("Events").document(eventId).collection("signups")
+                .get()
+                .addOnSuccessListener(queryDocumentSnapshots -> {
+
+                    long signupsCount = queryDocumentSnapshots.size();
+                    Log.d(TAG, "signupsCount = "+signupsCount);
+
+                    // Check if conditions for sending notifications are met
+                    if (signupsCount == 1) {
+                        // Create a notification message based on the signups count
+                        String notificationMessage = "Milestone: Signups count for event '" + myeventTitle + "' is " + signupsCount + ".";
+                        // Send the notification to the organizer
+                        sendPushNotification(orgfcmToken, notificationMessage, eventId);
+//                        Toast.makeText(, "Milestone sent", Toast.LENGTH_SHORT).show();
+                    }
+                })
+                .addOnFailureListener(e -> {
+                    Log.e(TAG, "Failed to retrieve signups count", e);
+//                    Toast.makeText(new AnnouncementsAndMilestones(), "Failed to retrieve signups count", Toast.LENGTH_SHORT).show();
+                });
+    }
+
+
+    /**
+     * Sends a push notification to the specified FCM token with the given message.
+     *
+     * @param fcmToken    The FCM token of the device.
+     * @param messageText The message to be sent.
+     */
+    private void sendPushNotification(String fcmToken, String messageText, String eventId) {
+
+        new Thread(() -> {
+            OkHttpClient client = new OkHttpClient();
+            MediaType mediaType = MediaType.parse("application/json");
+            JSONObject jsonNotif = new JSONObject();
+            JSONObject wholeObj = new JSONObject();
+            try {
+                jsonNotif.put("title", "ANNOUNCEMENT - " + myeventTitle);
+                jsonNotif.put("body", messageText);
+                wholeObj.put("to", fcmToken);
+                wholeObj.put("notification", jsonNotif);
+            } catch (JSONException e) {
+                Log.d(TAG, e.toString());
+            }
+            RequestBody rBody = RequestBody.create(mediaType, wholeObj.toString());
+            Request request = new Request.Builder().url("https://fcm.googleapis.com/fcm/send")
+                    .post(rBody)
+                    .addHeader("Authorization", "key=AAAAlidLyZE:APA91bF1suXeK0OgT_eZIP9uEPOCarD4zMfUyLWvo5-ljaXKQp4wuZhU2Ik2C63QLZKsvKGnOuzNIh_56WCIl1R8-rENZFlPPrwAB8Corgtnba5w8pMpknuhzp7_q1dTyshB37uTu4EN")
+                    .addHeader("Content-Type", "application/json").build();
+            Log.d(TAG, "think working");
+//            Toast.makeText(getActivity(), "Announcement sent successfully", Toast.LENGTH_SHORT).show();
+
+            try {
+                // Execute the request synchronously
+                Response response = client.newCall(request).execute();
+                // Check if the notification was sent successfully
+                if (response.isSuccessful()) {
+                    // Call the helper method to store the notification data
+                    NotificationStorer.storeNotification(fcmToken, eventId, messageText);
+                    Log.d(TAG, "Notification sent successfully");
+                } else {
+                    // Handle the case where notification sending failed
+                    Log.e(TAG, "Failed to send notification: " + response.message());
+                }
+            } catch (IOException e) {
+                Log.e(TAG, "IOException: " + e.getMessage());
+            }
+        }).start();
     }
